Migrate to <PackageReference> and new .csproj format.
This is step 1 in a .NET Core migraton, doing pieces first. However, building from the command line is far more of a pain than it should be. Uploading the current state as a checkpoint for what a user runs into in the real world.

Challenege: make the simplest build.bat or build.ps1 that builds this project from the CLI. Hopefully with a maximum of only the VS 2017 build tools installed (*not* a full Visual Studio install, which is unreasnable for most build servers).

For the .bat/ps1 file:
A user should be able:
1. Install VS 2017 build tools
2. Pull code
3. Run build.bat/ps1